refactor: extract ui widget helpers
